Data Science in a Box: Denny and LaQuinta Analysis
This analysis attempts to provide a better answer to the problem of if La Quintas is closer to Denny's than the previous labs we have completed. In order to improve on these labs, we have selected other random restaurants to compare Dennys to to see if in fact Denny's is close to La Quintas or if restaurants in general just happen to be placed close to La Quintas.